Update index.html
Commented out some of the posts on the home page. I think we need to get those Jekyl posts on the page so its not hard coded but im not sure how to do that therefore I commented them out not deleted them. Also, changed the small paragraph on the index page introducing the blog.